refactor: move types fix vercel deploy
