+++ src/api/helpers/requestWrapper.ts
@@ -1,16 +1,13 @@
 import type { AxiosRequestConfig } from 'axios';
 
 import { API_URL } from '~api/constants/API_URL';
-import type { ApiAxiosResponse } from '~api/types/backend/responses/ApiAxiosResponse';
+import type { ApiResponse } from '~api/types/responses/ApiResponse';
 import { isClient } from '~helpers/isClient';
 import { returnAppError } from '~helpers/returnAppError';
 
 export const requestWrapper = async <T>(
-	callbackReq: (
-		apiUrl: string,
-		requestConfig?: AxiosRequestConfig,
-	) => Promise<ApiAxiosResponse<T>>,
-): Promise<ApiAxiosResponse<T>> => {
+	callbackReq: (apiUrl: string, requestConfig?: AxiosRequestConfig) => Promise<ApiResponse<T>>,
+): Promise<ApiResponse<T>> => {
 	const apiUrl = isClient() ? API_URL.client : API_URL.server;
 
 	try {
