+++ src/api/types/responses/ApiResponse.ts
@@ -1,4 +1,4 @@
 import type { ErrorResponse } from './ErrorResponse';
 import type { SuccessResponse } from './SuccessResponse';
 
-export type ApiAxiosResponse<T> = SuccessResponse<T> | ErrorResponse;
+export type ApiResponse<T> = SuccessResponse<T> | ErrorResponse;
