+++ src/api/types/responses/ErrorResponse.ts
