+++ src/apiAuth/common/postAuthCommon.ts
@@ -1,12 +1,9 @@
 import axios from 'axios';
 
-import type { ApiAxiosResponse } from '~api/types/backend/responses/ApiAxiosResponse';
+import type { ApiResponse } from '~api/types/responses/ApiResponse';
 import { returnAppError } from '~helpers/returnAppError';
 
-export const postAuthCommon = async <R, B>(
-	endpoint: string,
-	body: B,
-): Promise<ApiAxiosResponse<R>> => {
+export const postAuthCommon = async <R, B>(endpoint: string, body: B): Promise<ApiResponse<R>> => {
 	try {
 		const { data, status } = await axios.post(endpoint, body);
 		return { data, status };
