+++ src/apiAuth/rest/auth.ts
@@ -1,4 +1,4 @@
-import type { ApiAxiosResponse } from '~api/types/backend/responses/ApiAxiosResponse';
+import type { ApiResponse } from '~api/types/responses/ApiResponse';
 
 import { getAuthCommon } from '../common/getAuthCommon';
 import { postAuthCommon } from '../common/postAuthCommon';
@@ -12,9 +12,9 @@ interface AuthAdminResponse {
 	token: string;
 }
 
-export const login = async (body: AuthAdmin): Promise<ApiAxiosResponse<AuthAdminResponse>> =>
+export const login = async (body: AuthAdmin): Promise<ApiResponse<AuthAdminResponse>> =>
 	postAuthCommon('/api/auth/login', body);
 
 // TODO: replace get to correct delete or patch
-export const logout = async (): Promise<ApiAxiosResponse<AuthAdminResponse>> =>
+export const logout = async (): Promise<ApiResponse<AuthAdminResponse>> =>
 	getAuthCommon('/api/auth/logout');
