+++ src/pageComponents/AdminPage/Dashboard/ContentBoard/Boards/ListsBoard/ListsBoard.tsx
@@ -1,6 +1,6 @@
 import { useListsState } from '~/pageComponents/AdminPage/store/useListsState';
 import { useTabsState } from '~/pageComponents/AdminPage/store/useTabsState';
-import type { ListRequest } from '~api/types/backend/requests/ListRequest';
+import type { ListRequest } from '~api/types/requests/ListRequest';
 import { Loader } from '~components/Loader/Loader';
 
 import { ListTable } from './ListTable/ListTable';
